python_microservice/requirements.txt
django>=4.2
pymongo>=4.6
python-dotenv>=1.0
PyJWT>=2.6.0
djangorestframework>=3.14.0
gunicorn>=21.2